proper handling of the device status

+++ devauth/devauth.go
@@ -300,15 +300,12 @@ func (d *DevAuth) processPreAuthRequest(ctx context.Context, r *model.AuthReq) (
 	// check the device status
 	// if the device status is accepted then do not trigger provisioning workflow
 	// this needs to be checked before changing authentication set status
-	status, err := d.db.GetDeviceStatus(ctx, aset.DeviceId)
+	dev, err := d.db.GetDeviceById(ctx, aset.DeviceId)
 	if err != nil {
-		if err == store.ErrDevNotFound {
-			return nil, err
-		}
-		return nil, errors.Wrap(err, "Cannot determine device status")
+		return nil, err
 	}
 
-	if status == model.DevStatusAccepted {
+	if dev.Status == model.DevStatusAccepted {
 		deviceAlreadyAccepted = true
 	}
 
@@ -346,18 +343,38 @@ func (d *DevAuth) processPreAuthRequest(ctx context.Context, r *model.AuthReq) (
 		return nil, errors.Wrap(err, "failed to update auth set status")
 	}
 
-	if err := d.db.UpdateDevice(ctx,
-		model.Device{
-			Id: aset.DeviceId,
-		},
-		model.DeviceUpdate{
-			Status: model.DevStatusAccepted,
-		}); err != nil {
-		return nil, errors.Wrap(err, "failed to update auth set status")
+	if err := d.updateDeviceStatus(ctx, aset.DeviceId, model.DevStatusAccepted); err != nil {
+		return nil, err
 	}
 
 	aset.Status = model.DevStatusAccepted
 	return aset, nil
+}
+
+func (d *DevAuth) updateDeviceStatus(ctx context.Context, devId, status string) error {
+	if status == "" {
+		newStatus, err := d.db.GetDeviceStatus(ctx, devId)
+		switch err {
+		case nil:
+			status = newStatus
+		case store.ErrAuthSetNotFound:
+			status = model.DevStatusRejected
+		default:
+			return errors.Wrap(err, "Cannot determine device status")
+		}
+	}
+
+	if err := d.db.UpdateDevice(ctx,
+		model.Device{
+			Id: devId,
+		},
+		model.DeviceUpdate{
+			Status:    status,
+			UpdatedTs: uto.TimePtr(time.Now().UTC()),
+		}); err != nil {
+		return errors.Wrap(err, "failed to update device status")
+	}
+	return nil
 }
 
 // processAuthRequest will process incoming auth request and record authentication
@@ -506,7 +523,7 @@ func (d *DevAuth) DeleteAuthSet(ctx context.Context, devId string, authId string
 		return d.db.DeleteDevice(ctx, devId)
 	}
 
-	return nil
+	return d.updateDeviceStatus(ctx, devId, "")
 }
 
 func (d *DevAuth) AcceptDeviceAuth(ctx context.Context, device_id string, auth_id string) error {
@@ -525,15 +542,12 @@ func (d *DevAuth) AcceptDeviceAuth(ctx context.Context, device_id string, auth_i
 	// check the device status
 	// if the device status is accepted then do not trigger provisioning workflow
 	// this needs to be checked before changing authentication set status
-	status, err := d.db.GetDeviceStatus(ctx, device_id)
+	dev, err := d.db.GetDeviceById(ctx, device_id)
 	if err != nil {
-		if err == store.ErrDevNotFound {
-			return err
-		}
-		return errors.Wrap(err, "Cannot determine device status")
+		return err
 	}
 
-	if status == model.DevStatusAccepted {
+	if dev.Status == model.DevStatusAccepted {
 		deviceAlreadyAccepted = true
 	}
 
@@ -592,7 +606,11 @@ func (d *DevAuth) setAuthSetStatus(ctx context.Context, device_id string, auth_i
 		return ErrDevIdAuthIdMismatch
 	}
 
-	if status == model.DevStatusRejected || status == model.DevStatusPending {
+	if aset.Status == status {
+		return nil
+	}
+
+	if aset.Status == model.DevStatusAccepted && (status == model.DevStatusRejected || status == model.DevStatusPending) {
 		// delete device token
 		err := d.db.DeleteTokenByDevId(ctx, aset.DeviceId)
 		if err != nil && err != store.ErrTokenNotFound {
@@ -624,7 +642,11 @@ func (d *DevAuth) setAuthSetStatus(ctx context.Context, device_id string, auth_i
 		return errors.Wrap(err, "db update device auth set error")
 	}
 
-	return nil
+	if status == model.DevStatusAccepted {
+		return d.updateDeviceStatus(ctx, device_id, status)
+	} else {
+		return d.updateDeviceStatus(ctx, device_id, "")
+	}
 }
 
 func (d *DevAuth) RejectDeviceAuth(ctx context.Context, device_id string, auth_id string) error {
@@ -891,14 +913,14 @@ func (d *DevAuth) GetTenantDeviceStatus(ctx context.Context, tenantId, deviceId 
 		Tenant: tenantId,
 	})
 
-	status, err := d.db.GetDeviceStatus(tenantCtx, deviceId)
+	dev, err := d.db.GetDeviceById(tenantCtx, deviceId)
 	switch err {
 	case nil:
-		return &model.Status{Status: status}, nil
+		return &model.Status{Status: dev.Status}, nil
 	case store.ErrDevNotFound:
 		return nil, ErrDeviceNotFound
 	default:
-		return nil, errors.Wrapf(err, "cannot get status for device %s", deviceId)
+		return nil, errors.Wrapf(err, "get device %s failed", deviceId)
 
 	}
 }
